Implement the account verification system
It required making some changes to the Verification model, as well
as improving the path name of the routes, such that they appear as
more intuitive.
The VerificationController is responsible for starting the verification
process of a logged-in account, and ends it by checking if the code
that the user entered is valid (exists, correct, and not expired).
If valid, user is verified.
Currently, the controller is restricted to handling customer accounts.